+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * functions mount, mountStatic
 * each, index variable is not included by default
 * reference to element is removed on destroying
+* config.useGroupReferencing
 
 ## 0.6.x